fixing map and boxplot issue again

+++ plotwipe.py
@@ -3,7 +3,8 @@ folder = os.path.join('utakeout','static','img','plot')
 delete_count = 0
 for file in os.listdir(folder):
 	file_path = os.path.join(folder,file)
-	if os.path.isfile(file_path):
+	name, ext = os.path.splitext(file_path)
+	if os.path.isfile(file_path) and ext == '.png':
 		os.remove(file_path)
 		print '{} was deleted.'.format(file_path)
 		delete_count += 1
